Add blueprint for blog routes.
