+++ application/templates/blog/index.html
